refactor 'enter' button from RoomModalOngoingEvent into RoomModal
