+++ src/components/templates/PartyMap/components/RoomModalOngoingEvent/RoomModalOngoingEvent.scss
@@ -26,8 +26,4 @@
     margin-bottom: 10px;
     white-space: pre-wrap;
   }
-
-  .room-entry-button {
-    width: 100%;
-  }
 }
